[v0.3.0] miscellaneous bug fixes
The development environment dependencies were different than the versions
available on the test environment. This fixes the debian dependencies so
they match the expected versions and fixes errors resulting from the change
of version.

Additionally, the ArchiveRunner was throwing an error when instantiated by
a non-root user, because the logger was trying to create the log file at the
time of init. The fix for this is to not create the logger until first use,
after the permissions checks have already been done.